Fix warnings.

+++ source/base/codec/multi_channel_resampler.cc
@@ -53,7 +53,7 @@ MultiChannelResampler::MultiChannelResampler(int channels,
     // |destination| provided to ProvideInput() directly.
     if (channels > 1)
     {
-        resampler_audio_bus_ = AudioBus::Create(channels - 1, request_size);
+        resampler_audio_bus_ = AudioBus::Create(channels - 1, static_cast<int>(request_size));
         for (int i = 0; i < resampler_audio_bus_->channels(); ++i)
         {
             wrapped_resampler_audio_bus_->SetChannelData(i + 1, resampler_audio_bus_->channel(i));
